01-DocuMind-RAG/requirements.txt
streamlit
langchain
langchain-groq
langchain-huggingface
langchain-community
pypdf
chromadb
python-dotenv
sentence-transformers